phpstan

+++ Tests/Util/FilesTest.php
@@ -17,16 +17,18 @@ class FilesTest extends TestCase
         // set up
         mkdir($directory);
         mkdir($subDirectory);
-        $fp = fopen($file, 'w');
-        fwrite($fp, 'lorem');
-        fclose($fp);
 
-        // tests
-        $this->assertSame(true, Files::isDirectory($directory));
-        $this->assertSame(true, Files::isFile($file));
-        Files::deleteDirectory($directory);
-        $this->assertSame(false, Files::isDirectory($directory));
-        $this->assertSame(false, Files::isDirectory($subDirectory));
-        $this->assertSame(false, Files::isFile($file));
+        if ($fp = fopen($file, 'w')) {
+            fwrite($fp, 'lorem');
+            fclose($fp);
+
+            // tests
+            $this->assertSame(true, Files::isDirectory($directory));
+            $this->assertSame(true, Files::isFile($file));
+            Files::deleteDirectory($directory);
+            $this->assertSame(false, Files::isDirectory($directory));
+            $this->assertSame(false, Files::isDirectory($subDirectory));
+            $this->assertSame(false, Files::isFile($file));
+        }
     }
 }
